+++ packages/origin-ui/src/components/ProducingDeviceTable.tsx
@@ -72,7 +72,7 @@ interface IEnrichedProducingDeviceData {
 
 interface IProducingDeviceTableState extends IPaginatedLoaderFilteredState {
     detailViewForDeviceId: string;
-    showRequestIRECsModal: boolean;
+    showRequestCertificatesModal: boolean;
     paginatedData: IEnrichedProducingDeviceData[];
 }
 
@@ -83,7 +83,7 @@ class ProducingDeviceTableClass extends PaginatedLoaderFiltered<Props, IProducin
         this.state = {
             ...getInitialPaginatedLoaderFilteredState(),
             detailViewForDeviceId: null,
-            showRequestIRECsModal: false
+            showRequestCertificatesModal: false
         };
     }
 
@@ -120,7 +120,7 @@ class ProducingDeviceTableClass extends PaginatedLoaderFiltered<Props, IProducin
         });
     }
 
-    async requestIRECs(rowIndex: number) {
+    async requestCerts(rowIndex: number) {
         this.props.showRequestCertificatesModal({
             producingDevice: this.state.paginatedData[rowIndex].device
         });
@@ -240,8 +240,8 @@ class ProducingDeviceTableClass extends PaginatedLoaderFiltered<Props, IProducin
         ) {
             actions.push({
                 icon: <Assignment />,
-                name: 'Request I-RECs',
-                onClick: (row: number) => this.requestIRECs(row)
+                name: 'Request Certificates',
+                onClick: (row: number) => this.requestCerts(row)
             });
         }