+++ packages/origin-ui/src/features/contracts/reducer.ts
@@ -2,21 +2,16 @@ import { ContractsActions, IContractsAction } from './actions';
 
 export interface IContractsState {
     marketContractLookupAddress: string;
-    currencies: string[];
 }
 
 const defaultState: IContractsState = {
-    marketContractLookupAddress: '',
-    currencies: []
+    marketContractLookupAddress: ''
 };
 
 export default function reducer(state = defaultState, action: IContractsAction): IContractsState {
     switch (action.type) {
         case ContractsActions.setMarketContractLookupAddress:
             return { ...state, marketContractLookupAddress: action.payload.address };
-
-        case ContractsActions.setCurrencies:
-            return { ...state, currencies: action.payload.currencies };
 
         default:
             return state;