+++ packages/origin/src/test/CertificateLogic.test.ts
@@ -15,7 +15,7 @@ import {
     DeviceLogic,
     Contracts as DeviceRegistryContracts
 } from '@energyweb/device-registry';
-import { Configuration, Compliance } from '@energyweb/utils-general';
+import { Configuration } from '@energyweb/utils-general';
 import { OffChainDataClientMock } from '@energyweb/origin-backend-client';
 import { deployERC721TestReceiver } from './deploy';
 import { TestReceiver } from '../wrappedContracts/TestReceiver';
@@ -212,7 +212,7 @@ describe('CertificateLogic-Facade', () => {
             gpsLongitude: '99.977800',
             timezone: 'Asia/Bangkok',
             deviceType: 'Wind',
-            complianceRegistry: Compliance.EEC,
+            complianceRegistry: 'I-REC',
             otherGreenAttributes: '',
             typeOfPublicSupport: '',
             description: '',