+++ packages/market-matcher-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/energywebfoundation/origin/compare/@energyweb/market-matcher-core@1.1.0...@energyweb/market-matcher-core@1.1.1) (2019-12-17)
+
+**Note:** Version bump only for package @energyweb/market-matcher-core
+
+
+
+
+
 # [1.1.0](https://github.com/energywebfoundation/origin/compare/@energyweb/market-matcher-core@1.0.1...@energyweb/market-matcher-core@1.1.0) (2019-12-12)